+++ index.html
@@ -8,5 +8,7 @@
 			canvas { width: 100%; height: 100% }
 		</style>
   </head>
-  <body></body>
+  <body>
+    <div id="root"></div>
+  </body>
 </html>
