+++ package.json
@@ -4,10 +4,14 @@
   "description": "Randomly generated solar system in Three.js",
   "private": true,
   "dependencies": {
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "three": "^0.110.0"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
+    "@babel/core": "^7.7.4",
+    "@babel/preset-env": "^7.7.4",
+    "@babel/preset-react": "^7.7.4",
     "babel-loader": "^8.0.6",
     "file-loader": "^4.3.0",
     "html-webpack-plugin": "^3.2.0",
